IT Department Budget vs Spend dashboard in PowerBI
This dashboard was created using 12 months of budget and forecast data and 6 months of actual data for the IT apartment of a global company. This is a mock dataset used to demonstrate my understanding of Power BI.

The budget, forecast and actuals have been stored in CSV files that have been exported from an accounting system.

The files contain records of date, department, cost element, control and the amount. 

This dashboard compares performance against budget