IT Department Budget vs Spend dashboard in Power BI
Interactice Dashboard 

https://app.powerbi.com/view?r=eyJrIjoiNGYyYjU4ZmMtZjhjZC00ZWE3LTlmOTgtMGZjZmE2M2Y0ZjQ4IiwidCI6ImQyZmM1ZGFhLWVjMDEtNDUzNS05YmYyLTg3OTk5MjM4Nzk1YSJ9


This dashboard was created using 12 months of budget and forecast data and 6 months of actual data for the IT apartment of a global company. This is a mock dataset used to demonstrate my understanding of Power BI.

The budget, forecast and actuals have been stored in CSV files that have been exported from an accounting system.

The files contain records of date, department, cost element, control and the amount. 

This dashboard compares performance against budget.